Cast the number.size() value to the integer type value in the print function.

+++ src/functions.cpp
@@ -4,7 +4,7 @@
 #include <limits>
 
 void print(std::vector<int> const & numbers) {
-    for (int i = 0; i < numbers.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(numbers.size()); ++i) {
         std::cout << i << ": " << numbers[i] << '\n';
     }
     std::cout << '\n';
